refactor: split game map area state logic and build stack (#3878)
Extract work-target, draft-projection, and province-action helpers from
game_map_area_state_logic.dart into library parts; split map canvas stack
from build overlays for Phase 3 flame map modularization.
