+++ SPEC/program/repo-and-packages.md
@@ -21,7 +21,7 @@ Monorepo layout:
 - **`colonizethis_app_e2e_support`** (`packages/colonizethis_app_e2e_support`) hosts E2E/widget mirror helpers (`e2e_helpers.dart`, `e2e_test_shared*.dart`, panel expected-line fixtures) previously under `app/integration_test/` and `app/test/e2e_*`. `app/test` retains only barrel/contract tests (≤10 `e2e_*.dart` files); integration scenarios import the support package. Seed-42 demo fixtures remain in `app/lib/test_support/` (production/widgetbook consumers).
 - **`features/game/flame/caches/`** — shared Flame asset decode helper (`AssetImageCache`) and map marker icon caches (town, port, fleet, civilian, resource, province label). Public barrel: `caches/caches.dart`.
 - **`features/game/flame/map_area/`** — `GameMapAreaBackground` and other render-only map chrome. Public barrel: `map_area/map_area.dart`.
-- **`features/game/flame/map_state/`** — `GameMapArea` widget stack, selection, province actions, turn-feed hooks, and draft projections. `game_map_area.dart` library parts: `game_map_area_build_overlays.dart` (map canvas stack + in-map overlay chrome), `game_map_area_build.dart` (controls bar + play-area shell). Public barrel: `map_state/map_state.dart`.
+- **`features/game/flame/map_state/`** — `GameMapArea` widget stack, selection, province actions, turn-feed hooks, and draft projections. `game_map_area.dart` library parts: `game_map_area_build_map_stack.dart` (map canvas stack + in-map overlay chrome), `game_map_area_build_overlays.dart` (play-area shell + debug/narrow slots), `game_map_area_build.dart` (controls bar + play-area shell). `game_map_area_state_logic.dart` library parts: `game_map_area_state_logic_work_targets.dart`, `game_map_area_state_logic_draft_projection.dart`, `game_map_area_state_logic_province_actions.dart`. Public barrel: `map_state/map_state.dart`.
 - **`features/game/flame/region_map/`** — `RegionMapComponent` Flame render stack and viewport helpers. `region_map_component.dart` library parts split shared constants/plate layout (`region_map_component_shared.dart`) from visibility modes (`region_map_component_shared_visibility.dart`), fog/transport helpers (`region_map_component_shared_visibility_fog_transport.dart`), extraction disc helpers (`region_map_component_shared_visibility_extraction.dart`), and label/halos (`region_map_component_shared_visibility_*.dart`); civilian vs fleet unit marker paint (`region_map_component_render_markers_units_civilian.dart`, `region_map_component_render_markers_units_fleet.dart`); capital rings (`region_map_component_render_markers_settlements_capitals.dart`), town/port paint (`region_map_component_render_markers_settlements_towns.dart`), and warp-zone edges (`region_map_component_render_markers_settlements_warp.dart`); province label compute vs paint (`region_map_component_render_political_labels_province_compute.dart`, `region_map_component_render_political_labels_province_paint.dart`). Public barrel: `region_map/region_map.dart`.
 - **`features/game/flame/tilesets/`** — terrain Wang tilesets, transport overlay tilesets, and connectivity masks. `terrain_tileset.dart` library parts: `terrain_tileset_models.dart` (Wang tile types, layer enum), `terrain_tileset_variant_keys.dart` (L2 variant tile-key resolution). Public barrel: `tilesets/tilesets.dart`.
 - **`features/game/flame/controls/`** — in-game map shell chrome: top/tab bar ([GameMapControls](../../app/lib/features/game/flame/controls/game_map_controls.dart)), corner tool row ([GameMapCornerControls](../../app/lib/features/game/flame/controls/game_map_corner_controls.dart)), and empire left rail ([GameMapEmpireLeftRail](../../app/lib/features/game/flame/controls/game_map_empire_left_rail.dart)). Public barrel: `controls/controls.dart`.
